+++ scripts/highlight.js
@@ -6,12 +6,15 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
 	if (!loaded) {
 		loaded = true;
 
+		console.log("here");
+
 		const functionDelay = request.functionDelay;
 		const values = request.values;
 		const unwantedTags = request.unwantedTags;
 		highlightingClass = request.highlightingClass
 
 		if (functionDelay && values && unwantedTags && highlightingClass) {	
+			console.log("here2");
 			const textChildNodes = obj => Array.from(obj.childNodes)
 				.filter(node => node.nodeName === "#text");
 
@@ -37,6 +40,7 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
 					//node.data = '';
 					// insert new child after the old child
 					//node.after(fragment);
+					console.log("highlighted");
 					return !matches ? 0 : matches.length;
 				}
 			}
@@ -56,11 +60,15 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
 				// check if there is any character to be highlighted
 				const nodesToBeHighlighted = allTags.filter(tag => {
 					const test = tag.textContent.match(kanjiRegex);
+					//console.log(kanjiRegex);
 					return test !== null ? test.length > 0 : false;
 				});
 
+				console.log("inside highlighter");
+
 				let nmrHighlightedKanji = 0;
 				if (nodesToBeHighlighted.length > 0) {
+					console.log(nodesToBeHighlighted);
 					const span = document.createElement("span");
 					span.className = className;
 					nodesToBeHighlighted.filter(tag => tagFilteringConditions(tag)).forEach(node => nmrHighlightedKanji += replaceMatchesWithElem(node, kanjiRegex, span));